scheduler/requirements.txt
alembic==1.8.1
APScheduler==3.9.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
Flask==2.1.3
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
pika==1.3.0
psycopg2==2.9.3
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.39
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.11
Werkzeug==2.2.1
zipp==3.8.1
